// enums.ts
export enum SpaceTypes {
  PROPERTY = "property",
  UTILITY = "utility",
  RAILROAD = "railroad",
  CHANCE = "chance",
  COMMUNITY_CHEST = "community_chest",
  FREE_PARKING = "free_parking",
  GO_TO_JAIL = "go_to_jail",
  INCOME_TAX = "income_tax",
  LUXURY_TAX = "luxury_tax",
  GO = "go",
}

export enum PropertyGroups {
  BROWN = "brown",
  LIGHT_BLUE = "light_blue",
  PINK = "pink",
  ORANGE = "orange",
  RED = "red",
  YELLOW = "yellow",
  GREEN = "green",
  DARK_BLUE = "dark_blue",
}